live_agent/requirements.txt
Flask 
flask-sock
assemblyai
python-dotenv
ngrok
twilio
deepgram-sdk 